Bugfix: wrong ID number Data-type

+++ src/ActionUnit.h
@@ -52,7 +52,7 @@ public:
     bool                  registerAction( TAction * pT );
     void                  unregisterAction( TAction * pT );
     void                  reParentAction( int childID, int oldParentID, int newParentID, int parentPostion = -1, int childPosition = -1 );
-    qint64                getNewID();
+    int                   getNewID();
     void                  uninstall(const QString& );
     void                  _uninstall( TAction * pChild, const QString& packageName );
     void                  updateToolbar();
@@ -79,7 +79,7 @@ private:
     QPointer<Host>        mpHost;
     QMap<int, TAction *>  mActionMap;
     std::list<TAction *>  mActionRootNodeList;
-    qint64                mMaxID;
+    int                   mMaxID;
     TToolBar *            mpToolBar;
     TEasyButtonBar *      mpEasyButtonBar;
     bool                  mModuleMember;
