+++ src/TTimer.cpp
@@ -316,7 +316,7 @@ bool TTimer::canBeUnlocked( TTimer * pChild )
 
 }
 
-void TTimer::enableTimer( qint64 id )
+void TTimer::enableTimer( int id )
 {
 
     if( mID == id )
@@ -351,7 +351,7 @@ void TTimer::enableTimer( qint64 id )
     }
 }
 
-void TTimer::disableTimer( qint64 id )
+void TTimer::disableTimer( int id )
 {
     if( mID == id )
     {
